KUriFilter: simplify data structures, fix memory leak
Summary:
This code had a QHash and a separate ordered list of names in order
to use plugins in the right order. It's much simpler to just have a
vector of plugins instead. Also, this fixes a memory leak when multiple
plugins are found with the same name (e.g. one from /usr and one from
my custom prefix, although the right fix for this particular
issue is to load only one of them).

Test Plan: kurifilter tests still pass

Reviewers: apol, davidedmundson, arichardson, bshah

Reviewed By: apol

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9100