Fix crash, presumably since Qt 5.10?
Summary:
We are opening the stream as read only and skipRawData would complain that
it's a write-only device, and skipping needs reading:
QIODevice::skip (QBuffer): WriteOnly device
It was a weird optimization anyway, so just pass the information like the
rest of the data and it works just fine.

BUG: 386364

Test Plan:
The check on the bug report passes now, and akregator fetches
all feeds again.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: anthonyfieroni, ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9399