[KFileItemActions] Allow specifying the number of selected files required for an action
This was mostly prompted by D9698 since Kompare can only compare two files and it makes no sense to offer this option
for less or more than that.
It is done as a QStringList so you could specify multiple criteria, e.g. only show for 2 or 4 files.
If the need should arise, this could be made more sophisticated by adding support for e.g. ranges and less than/more
than operators, but let's keep it simple for now.

CHANGELOG: KonqPopupMenu/Plugin can now use the X-KDE-RequiredNumberOfUrls key to require a certain number of files to be selected before being shown

Differential Revision: https://phabricator.kde.org/D9734