Add unit tests for privilege excution in KIO
Summary:
Auto testing privilege operation is not feasible because it requires user interaction. So instead what we do
is check if the flag for privilege execution is set and if the sub-jobs have a parent job.
If the conditions are met then metadata "TestData" is added which tells that the job is properly configured
and can perform file operations with elevated privileges.

Differential Revision: https://phabricator.kde.org/D7430