Change an assert to do as the function documentation tells
Summary:
Instead of an assert do as the documentation tells: return an invalid Url.
* In case both URLs are equal, an invalid URL is returned

Test Plan:
How to assert:
In dolphin, in the location bar, change it into text and enter into one level directory from root, for example /d
When it changes into the breadcrum, press into the root one and crash (in Debug build).

I've been using this patch one month. And it doesn't affect the unit tests.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10071