Use the destination KFileItem object to figure out of it's writable instead of creating a KFileItemListProperties which basically does the same thing but more expensive.
Summary:
Use KFileItem::isWritable instead of the detour with KFileItemListProperties.
It's basically the same, only more expensive.
KFileItem::isWritable is already quite expensive on it's own.
This saves cycles but to be fair, the amount is neglectable.

Test Plan:
I verified that the paste action in Dolphin (under the Right mouse button) still works.
Is active (when writable) and shows the correct text as before.

Reviewers: dfaure, ngraham

Reviewed By: ngraham

Subscribers: ngraham, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10282