Use F11 as the shortcut to toggle the aside preview
Summary:
FEATURE: 389880
FIXED-IN: 5.44

Test Plan:
{F5692446}

Shortcut works and toggles the aside preview.

However, in doing so, another pre-existing bug is triggered: the file dialog doesn't eat the keyboard input, so if the app that spawned it has something bound to F11, you'll erroneously get the "Ambiguous shortcut" message (even though the keystroke is/should be passed only to the file dialog, not the host app). Help solving this would be appreciated.

Reviewers: #frameworks, markg

Reviewed By: markg

Subscribers: markg

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10314