Repair copying file to VFAT without warnings.
Summary:
The refactoring returned the wrong value, which led to warnings,
but it was also in the wrong place. We want KIO::chown() to fail
explicitly, while we want chown during a file copy to be ignored
on e.g. VFAT systems.

Test Plan: Copying a file to a VFAT partition no longer warns/errors.

Reviewers: chinmoyr, markg, jtamate

Reviewed By: jtamate

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10358