kio_file: skip error handling for initial perms during file copy
Summary:
This reverts the parts of 13ab6a85bb which touched this code, there
didn't use to be any error handling for this call, on purpose.
This is just the initial permissions for the newly created file we are
going to copy to; if the filesystem doesn't support permissions, we can
still proceed.

Test Plan: Copying a file to a VFAT partition no longer leads to an empty file.

Reviewers: chinmoyr, markg, jtamate

Reviewed By: jtamate

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10376