Fix "ambiguous shortcut" issue introduced with D10314
Summary:
My patch D10314 was insufficiently tested (sorry about that!) and introduced an "Ambiguous Shortcut" warning, because "show/hide inline previews" was already bound to F11.

This patch solves that issue by changing the shortcut to F12. F11 was never appropriate for it in the first place, since that's used for something different in Dolphin anyway. Dolphin doesn't currently have a shortcut for "show hide inline previews", so there's nothing to be consistent with. I used F12 in the file dialogs since it's unused in both places, so we could add it to Dolphin too if we really wanted.

Test Plan: Tested with Kate. F11 toggles the aside preview, and F12 toggles the inline preview, like they should. No more conflicts!

Reviewers: #frameworks, #dolphin, markg, elvisangelaccio

Reviewed By: markg

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10617