[KFileWidget] Hide places frame and header
This avoids a double "Places" header and also gets rid of the superfluous frame around the widget, similar to Dolphin.
It has never been possible to detach the dock from the window anyway.
The close button is also removed as a side-effect of this but should a user really want to hide it,
this can still be done by pressing F9 or in the menu. Resizing the places list is still possible.

Differential Revision: https://phabricator.kde.org/D10325