FileUndoManager: don't delete non-existing local files
Summary:
After a CopyJob the FileUndoManager records the file that was copied.
If this file is deleted before the Undo operation is triggered, the
File UndoManager will still try to undo the copy by deleting it.

This patch fixes this issue by validating the recorded files with
`QFileInfo::exists()`. If the FileUndoManager realizes that is no longer
possible to undo the CopyJob, it will signal that the undo operation is
not available anymore.

Test Plan:
From Dolphin:
- Copy foo.txt to bar.txt
- Shift+Del bar.txt
- CTRL+Z

Reviewers: dfaure

Subscribers: ngraham, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10312