Don't try to find an icon for an empty url
Summary:
The message kf5.kio.core: Refilling KProtocolInfoFactory cache in the hope to find ""
was produced when findProtocol was called by KProtocolInfo::icon that was
called with an empty url.

Test Plan: No more Refilling KProtocolInfoFactory cache in the hope to find "" when browsing net or samba shares.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: broulik

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11044