[KFilePlacesView] Call KDiskFreeSpaceInfo only if neccessary
I found this was called 350 times for me when opening the file dialog since for every paint pass it would needlessly call this for every place.
It queries all mountpoints and does other expensive work.
Instead, only do so when the bar would actually be shown. Ideally this would be cached somehow but this patch is a nice simple optimization already.

Differential Revision: https://phabricator.kde.org/D11088