Don't show top "Open With" app for folders; only for files
Summary:
This patch returns the {nav Open With}  items for folders to the way it was before D11569: with all entries shown in a sub-menu. This should make the context menu more relevant since most of the time when you right click on a folder, it's not to open it in some app.

The top app is still displayed inline for the app context menu--just not for folders.

Test Plan:
Top app {nav Open With} entries...
- Do not appear for local and remote folders and symlinks to folders
- Appear for local and remote files and symlinks to files

Folder:
{F5807966}

File:
{F5807934}

Note: screenshots reflect the {nav Open With} positioning from {D11884}

Reviewers: #frameworks, elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: rkflx, elvisangelaccio, Zren, michaelh

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D12206