KFileWidget: Set minimum size for zoom slider
Summary:
When choosing a very narrow width for the file dialog, the zoom slider
for the icon size would get squished so that only the handle would be
visible and the slider could not be used anymore to change zoom levels
in any way.

This can be avoided by setting a minimum size. 40px are smaller than the
default size, but still allow to operate the slider comfortably.

The minimal width of the dialog is unchanged, as before the toolbar will
show the elision button once elements start overflowing.

Ref T8552

FIXED-IN: 5.46

Test Plan:
`kdialog --getopenfilename`, change window size and observe zoom slider.

{F5826728}

Reviewers: #frameworks, ngraham

Reviewed By: ngraham

Subscribers: ngraham

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D12588