Use a regular dialog rather than a list dialog when trashing or deleting a single file
Summary:
This patch uses a regular-style dialog instead of a list-style dialog when the user is trying to trash or delete only a single item. It looks a bit nicer than showing a dialog that has a list with only a single item in it.

Also change the string "Trash" to "Move to Trash", to prevent ambiguity in languages where the word can be either a noun or a verb.

Depends on D12420

Test Plan:
Made sure that trash and deletion operations all still work. Here's how it looks:

Delete a single file:
{F5832555}

Delete multiple files:
{F5832429}

Trash a single file:
{F5832556}

Trash multiple files:
{F5832436}

Reviewers: #frameworks, #dolphin, elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: elvisangelaccio, cfeck

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D12422