Improve grid icon layout in filepicker dialog
Summary:
Icon grid layout now changes depending on the dialog window size.

As a side effect, this patch improves the grid spacing in icons-on-top mode by making it looser for small icons (which gives the labels more space) and tighter for large icons (which allows more to be seen at once).

BUG: 334099
Fixed-IN: 5.46

Test Plan:
Before:

{F5811621}

After:

{F5811623}
{F5811625}

Works just as well when scaling is used:

{F5811628}

Reviewers: #frameworks, #vdg, ngraham, rkflx

Reviewed By: #vdg, ngraham, rkflx

Subscribers: abetts, rkflx, ngraham, #frameworks

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D12306