KFileWidget: Disable zoom buttons once reached minimum or maximum
Summary:
In Gwenview the zoom buttons will be disabled when clicking on it would
not result in any further change of zoom level, e.g. {nav Zoom In} is
disabled for the largest icon size. In general it is recommended to
disable unavailable actions.

This patch adds the functionality to the file dialog too.

Ref T8552

FIXED-IN: 5.46

Test Plan:
`kdialog --getopenfilename`, change icon size to minimum and maximum by
sliding, mouse wheeling, pressing buttons and using the keyboard, and
check that buttons are disabling and enabling properly.

{F5826731}

Reviewers: #frameworks, ngraham

Reviewed By: ngraham

Subscribers: ngraham

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D12590