Thumbnail smooth scaling in filepicker
Summary:
The scaling of thumbnail previews in the Open/Save dialog was done with the fast transformation that looked out of place and ugly.
This uses the same scaling method as Dolphin to look consistent and nice too.

BUG: 345578
FIXED-IN: 5.46

Test Plan:
Before:

{F5815324}

After:

{F5815325}

Dolphin for reference uses smooth transformation:

{F5815326}

Reviewers: #frameworks, #vdg, ngraham, elvisangelaccio, rkflx

Reviewed By: #vdg, ngraham, rkflx

Subscribers: cfeck, broulik, ngraham, #frameworks

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D12385