Preserve fragment when redirecting from http to https
Summary:
Currently, a HTTP to HTTPS redirect removes the fragment if we have the following options set:
In Default Applications -> Web Browser
"Open http and https URLs"
to
"in an application based on the contents of the URL".

Running the command
`kde-open5 http://en.wikipedia.org/wiki/URL#Syntax`
sends the URL
`https://en.wikipedia.org/wiki/URL`
to the browser.

Reviewers: #plasma, dfaure

Reviewed By: dfaure

Subscribers: dfaure, michaelh, ngraham, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10903