Give the file dialogs a "Sort by" menu button on the toolbar
Summary:
This patch moves the sort order chooser and options out of the somewhat hidden Settings menu and onto the main toolbar.

Nice-to-haves that would make this more user-friendly in the future:
- Better "sort options are here" icon: https://bugs.kde.org/show_bug.cgi?id=393318.
- Breeze could draw a downward-pointing arrow for menu toolbuttons: https://bugs.kde.org/show_bug.cgi?id=344746

Test Plan:
New toolbar sorting menu:
{F5852072}

It's still there in the context menu and fully accessible for `KDirOperator` clients, but it's no longer in the file dialog's settings menu.

Reviewers: #frameworks, #dolphin, #vdg, rkflx

Reviewed By: rkflx

Subscribers: kde-frameworks-devel, andreaska, markg, broulik, anemeth

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D12337