Switch KIO::convertSize() to KFormat::formatByteSize()
Summary:
Make KIO::convertSize() use KFormat::formatByteSize(), instead of using
an own copy of the same function. The handling of the user
configuration of the default dialect is left as-is, and it should work
fine since the two enums BinaryUnitDialect (the removed one, and the
one in KFormat) are the same.

Test Plan: Builds fine, no changes to the result of KIO::convertSize().

Reviewers: dfaure, aacid

Reviewed By: aacid

Subscribers: aacid, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13549