Present error dialog when user tries to create directory named "." or ".."
Summary:
BUG: 387449

In Dolphin, when a user tries to create a folder named "." or "..", they are presented with a dialog that informs them that since the folder name they entered begins with a ".", the folder they are about to create will be hidden. This is, of course, misleading. This patch instead presents them with a similar dialog, informing them of the error and asking them for a new name.

Before:

{F5976307}

After:

{F5976315}

Reviewers: #dolphin, #frameworks, ngraham, dfaure, elvisangelaccio

Reviewed By: #dolphin, ngraham, elvisangelaccio

Subscribers: cfeck, elvisangelaccio, dfaure, tmarshall, bruns, ngraham, kde-frameworks-devel

Tags: #frameworks, #dolphin

Differential Revision: https://phabricator.kde.org/D13805