Top-align labels in properties dialog
Summary:
The {nav General} tab of KIO's properties dialog shows labels on the
left side, and properties on the right, creating a table-like structure.
Commonly a label should be aligned vertically in such a way it matches
the first line of the properties, to make it clear which properties it
belongs to.

However, this was not yet the case for the {nav Size} label.

Test Plan:
In Dolphin, select a folder and press {key Alt Return}. All labels in
{nav Properties > General} should be top-aligned.
Before: {F6165126}
After: {F6165125}

Reviewers: dhaumann, ngraham, cfeck

Reviewed By: ngraham, cfeck

Subscribers: cfeck, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14501