kioexecd: watch for creations or modifications of the temporary files
Summary:
When a non KIO friendly program opens a non local file, the file is copied to an user temporary folder by kioexec.
Watch any creation or modification or deletion of that temporary file, because some programs delete the old file and create a new one when saving the file. When the file is recreated or modified, the user is asked if he/she wants to upload the modified file.

If the temporary file keeps deleted more than 30s then forget about that file and try to delete the temporary directory.

BUG: 397742

Test Plan:
Tested logging out and in several times and opening two or three remote files with xed and libreoffice.
After removing manually some of the temporary files and wait around ~31s the empty temporary directories are deleted.

Reviewers: #frameworks, broulik, ngraham, dfaure, elvisangelaccio

Reviewed By: dfaure, elvisangelaccio

Subscribers: bruns, anthonyfieroni, elvisangelaccio, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15180