Improve "insufficient disk space" error message
Summary: The prior error was misleading, and always said "Disk full." even when the disk wasn't actually full; not having enough space to move or copy the requested file doesn't mean that the disk is full! This patch re-words the error message to be clearer and still work when `errorString` is defined (as it always should be).

Test Plan:
Before:
{F6268157}

After:
{F6268158}

Reviewers: #frameworks, #vdg, #dolphin, shubham

Reviewed By: shubham

Subscribers: broulik, abetts, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15557