Make it possible to change directory up even with trailing slashes in the url
Summary:
Normalize the paths in KUriFilterData.
Normalize the path in cdUp (leave only the last / in the url) before going up.

Test Plan:
The url of a ntfs partition is file:///d//. When I pressed cdUp, it remained in the same directory instead of going to root.
Now even if I write urls like

```
/d/c///////////////////////
```
when I press cdUp it goes up to /d/ and when I press Enter, the url stored in the history is /d/c/.

Reviewers: dfaure, #frameworks

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13676