Add the possibility to have a keyboard shortcut to create a file
Summary:
It is already possible to create a directory with a keyboard shortcut (F10 in
Dolphin). This patch adds the same possibility for creating a file.
The file template used will be the first one encountered in the list, so it should
be a text file by default, or a user defined one if any.

Reviewers: #frameworks, dfaure, elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: elvisangelaccio, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D16412