[KSambaShare] Trim trailing / from share path
Summary:
When adding usershares manually on the command line, it is common to
have a path with trailing "/" due to tab completion. Samba treats paths
with and without / the same, but KIO does not pick up these shares.

These folder paths are not marked in e.g. dolphin as shared and can not
be edited properly.

Test Plan:
make && ctest
open directory with shared folders in dolphin, also share with trailing / in `path=` show up

Reviewers: #frameworks, broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D16464