[ioslaves/trash] Handle broken symlinks in deleted subdirectories
Summary:
The trash ioslave currently handles:
 - working symlinks at the toplevel
 - broken symlinks at the toplevel
 - working symlinks in subdirectories

but broken symlinks in a subdirectory get lost.

This is because subdirs are populated from a QDir::listEntries in the
subdirectory folder, which by default filters out broken symlinks.

It's important to list them so that we we restore the directory as
otherwise we lose data. The symlink could become non-dead once files are
restored.

BUG: 400990

Test Plan: Relevant unit test

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17021