Restore sendfile support
Summary:
Somehow in the kdelibs -> framework port the cmake checks for
HAVE_SENDFILE got lost.

That re-enables a massive optimisation in the file kioslave that has all the code existing and used in kdelibs4 that we're currently missing.

Test Plan:
Put a compilation fail inside the #ifdef, before it wasn't triggered, now it is.

Ran unit tests
Moved a file in dolphin

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: ngraham, apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17048