Use correct variable type for returned value from read/sendfile
Summary:
n stores the read/transferred bytes. This returns a ssize_t.
We were casting to an int, which theoretically is a loss of data.

In practice it isn't an issue as we only read a max of MAX_IPC_SIZE at a
time, which would fit in an int.

Reviewers: apol

Reviewed By: apol

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17051