Fix assert when cleaning up symlinks in PreviewJob
Summary:
When previewing files in a KIOslave KIO will copy the file out to a
temporary directory and then run the preview on that.

If I have a symlink in a slave (for example trashing a symlink) the
preview job will copy the symlink to the temporary directory.

QFileInfo follows symlinks for methods like isDir(), in this case we
shouldn't assert as QFile::remove will correctly just cleanup the
symlink.

Test Plan:
Created a symlink to a directory
Trashed it
Opened trash:// in dolphin with previews enabled
Used to hit an assert
Now don't

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17017