[KUrlNavigator] Fix firstChildUrl when going back from archive
Summary:
When navigating in a archive file in Dolphin and then
using the URL navigator to go back to a parent folder,
the selection was wrong because we were calling
urlSelectionRequested with an URL that still had the
tar: scheme instead of file:.

Test Plan:
Navigate in a .zip archive within Dolphin. Click on any
parent folder, the selection should now be on the last
previously opened item (either a foler or the .zip file).

Reviewers: #frameworks, elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17256