[KDirOperator] Don't squeeze Name column when there isn't enough horizontal space
Summary:
When a `KDirOperator` view is used in {nav Detailed} or {nav Detailed Tree} view, when there isn't enough horizontal space, the {nav Name} column currently gets squeezed and becomes illegible.

This patch fixes that by making the name column fixed-width so it never gets too small to read the contents; instead the view sprouts a horizontal scrollbar, which is preferable.

BUG: 401506
FIXED-IN: 5.54

Test Plan:
In Kate's Filesystem Browser plugin sidebar view: {F6466450}

In the open/save dialog when there's not enough horizontal space: {F6466451}

In the open/save dialog when there //is// enough horizontal space: {F6466452}

In this case, the other columns no longer stick to the right size of the view when the window is resized, which maybe isn't as nice as current behavior. Open to suggestions.

Reviewers: #vdg, #frameworks, cullmann, cfeck

Reviewed By: cullmann, cfeck

Subscribers: cfeck, cullmann, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17461